Rename profile.go file as delete.go
